Start to split off formatting routines..
from builtin
